WARNING = CHANGED SO MANY THINGS hh/h/l/ll for int working + removing errors appearing when compiling with gcc -Werror

+++ LIBFT/ft_absolute.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_absolute(int i)
+#include <stdint.h>
+
+intmax_t ft_absolute(intmax_t i)
 {
 	if (i < 0)
 		return (-i);
